+++ tests/conftest.py
@@ -3,6 +3,8 @@ from unittest.mock import create_autospec
 import pytest
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.service import iam
+from pyspark.sql.types import StructType, StructField, LongType, StringType, TimestampType, IntegerType, BooleanType, \
+    ArrayType, MapType
 
 from databricks.labs.remorph.reconcile.recon_config import (
     Table,
@@ -11,7 +13,7 @@ from databricks.labs.remorph.reconcile.recon_config import (
     ColumnThresholds,
     Filters,
     TableThresholds,
-    ColumnMapping,
+    ColumnMapping, Schema,
 )
 
 
@@ -78,3 +80,121 @@ def table_conf():
         )
 
     return _table_conf
+
+
+@pytest.fixture
+def table_schema():
+    sch = [
+        Schema("s_suppkey", "number"),
+        Schema("s_name", "varchar"),
+        Schema("s_address", "varchar"),
+        Schema("s_nationkey", "number"),
+        Schema("s_phone", "varchar"),
+        Schema("s_acctbal", "number"),
+        Schema("s_comment", "varchar"),
+    ]
+
+    sch_with_alias = [
+        Schema("s_suppkey_t", "number"),
+        Schema("s_name", "varchar"),
+        Schema("s_address_t", "varchar"),
+        Schema("s_nationkey_t", "number"),
+        Schema("s_phone_t", "varchar"),
+        Schema("s_acctbal_t", "number"),
+        Schema("s_comment_t", "varchar"),
+    ]
+
+    return sch, sch_with_alias
+
+
+@pytest.fixture
+def report_tables_schema():
+    recon_schema = StructType(
+        [
+            StructField("recon_table_id", LongType(), nullable=False),
+            StructField("recon_id", StringType(), nullable=False),
+            StructField("source_type", StringType(), nullable=False),
+            StructField(
+                "source_table",
+                StructType(
+                    [
+                        StructField('catalog', StringType(), nullable=False),
+                        StructField('schema', StringType(), nullable=False),
+                        StructField('table_name', StringType(), nullable=False),
+                    ]
+                ),
+                nullable=False,
+            ),
+            StructField(
+                "target_table",
+                StructType(
+                    [
+                        StructField('catalog', StringType(), nullable=False),
+                        StructField('schema', StringType(), nullable=False),
+                        StructField('table_name', StringType(), nullable=False),
+                    ]
+                ),
+                nullable=False,
+            ),
+            StructField("report_type", StringType(), nullable=False),
+            StructField("operation_name", StringType(), nullable=False),
+            StructField("start_ts", TimestampType()),
+            StructField("end_ts", TimestampType()),
+        ]
+    )
+
+    metrics_schema = StructType(
+        [
+            StructField("recon_table_id", LongType(), nullable=False),
+            StructField(
+                "recon_metrics",
+                StructType(
+                    [
+                        StructField(
+                            "row_comparison",
+                            StructType(
+                                [
+                                    StructField("missing_in_source", IntegerType()),
+                                    StructField("missing_in_target", IntegerType()),
+                                ]
+                            ),
+                        ),
+                        StructField(
+                            "column_comparison",
+                            StructType(
+                                [
+                                    StructField("absolute_mismatch", IntegerType()),
+                                    StructField("threshold_mismatch", IntegerType()),
+                                    StructField("mismatch_columns", StringType()),
+                                ]
+                            ),
+                        ),
+                        StructField("schema_comparison", BooleanType()),
+                    ]
+                ),
+            ),
+            StructField(
+                "run_metrics",
+                StructType(
+                    [
+                        StructField("status", BooleanType(), nullable=False),
+                        StructField("run_by_user", StringType(), nullable=False),
+                        StructField("exception_message", StringType()),
+                    ]
+                ),
+            ),
+            StructField("inserted_ts", TimestampType(), nullable=False),
+        ]
+    )
+
+    details_schema = StructType(
+        [
+            StructField("recon_table_id", LongType(), nullable=False),
+            StructField("recon_type", StringType(), nullable=False),
+            StructField("status", BooleanType(), nullable=False),
+            StructField("data", ArrayType(MapType(StringType(), StringType())), nullable=False),
+            StructField("inserted_ts", TimestampType(), nullable=False),
+        ]
+    )
+
+    return recon_schema, metrics_schema, details_schema
